refactor: replace routes.ts with router.ts, adding travel plans route and authentication middleware
