--- backend/src/routers/routes.ts
@@ -1,27 +1,0 @@
-import cors from "cors";
-import express, {
-	type Express,
-	type Request,
-	type Response,
-	Router,
-} from "express";
-
-const setupRoutes = (app: Express) => {
-	app.use(express.json());
-	app.use(
-		cors({
-			origin: process.env.CORS_ORIGIN || "",
-			methods: ["GET", "POST", "OPTIONS"],
-		}),
-	);
-
-	const appRouter = Router();
-
-	appRouter.get("/", (_req: Request, res: Response) => {
-		res.json({ message: "OK" });
-	});
-
-	app.use("/api", appRouter);
-};
-
-export default setupRoutes;
